requirements-test.txt
# Test dependencies for AI Dev Flow
# Reference: ai_dev_flow/10_TSPEC/

# Core testing framework
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0

# Parallel execution (optional, uncomment when needed)
# pytest-xdist>=3.5.0

# JSON test reporting
pytest-json-report>=1.5.0

# YAML support for configuration and schemas
pyyaml>=6.0.1

# Schema validation
jsonschema>=4.21.0

# Optional: Enhanced testing features
# pytest-html>=4.1.0       # HTML reports
# pytest-bdd>=7.0.0        # BDD integration with Gherkin
# testcontainers>=3.7.0    # Docker test services
# pytest-asyncio>=0.23.0   # Async test support
# pytest-mock>=3.12.0      # Enhanced mocking

# Optional: Database testing
# pytest-postgresql>=5.0.0
# pytest-redis>=3.0.0

# Optional: API testing
# httpx>=0.27.0            # Async HTTP client for testing
# responses>=0.25.0        # Mock HTTP responses
